+++ CHANGELOG.md
@@ -12,6 +12,8 @@
     -   User bots no longer register their own context. Instead, a new bot has been created to host the `aux.users` context.
         -   Improves performance of AUXes with many user bots with the same username.
         -   Existing user bots are not affected. They will be deleted automatically if given enough time. Alternatively, you can delete them using `destroy(getBots("#aux._user"))`.
+-   Bug Fixes
+    -   Fixed issue where default panning tag locked the vertical movement in player.
 
 ## V0.11.0